+++ src/common/helpers/allureHelpers.js
@@ -11,7 +11,7 @@ export function parseTestTreeHierarchy(fileName, logger) {
     capitalize(camelCaseToPhrase(attribute)),
   );
 
-  if (attributes[2].includes('.spec.js')) {
+  if (attributes.length > 2 && attributes[2].includes('.spec.js')) {
     attributes = attributes.slice(0, 2);
   }
