+++ src/components/BaseComponent.js
@@ -1,0 +1,5 @@
+export class BaseComponent {
+  constructor(page) {
+    this.page = page;
+  }
+}
