+++ src/components/Cup.js
@@ -1,0 +1,31 @@
+import { BaseComponent } from './BaseComponent.js';
+import { expect, step } from '@playwright/test';
+
+export class Cup extends BaseComponent {
+  constructor(page) {
+    super(page);
+  }
+
+  coffeeCupLocator(coffeeName) {
+    const testId = coffeeName.replace(' ', '_');
+
+    return this.page.getByTestId(testId);
+  }
+
+  coffeeCupCostLocator(coffeeName) {
+    const coffeeCup = this.coffeeCupLocator(coffeeName);
+    return this.page.getByRole('listitem').filter({ has: coffeeCup });
+  }
+
+  async clickCoffeeCup(coffeeName) {
+    await step(`Click ${coffeeName} cup`, async () => {
+      await this.coffeeCupLocator(coffeeName).click();
+    });
+  }
+
+  async assertCoffeeCupCostHasValue(coffee, value) {
+    await step(`Assert ${coffee} cup cost has value: ${value}`, async () => {
+      await expect(this.coffeeCupCostLocator(coffee)).toContainText(value);
+    });
+  }
+}
