+++ src/pages/MenuPage.js
@@ -1,69 +1,17 @@
 import { expect, step } from '@playwright/test';
 import { BasePage } from './BasePage';
+import { Header } from '../components/Header';
+import { Promo } from '../components/Promo';
+import { Cup } from '../components/Cup';
+import { TotalCount } from '../components/TotalCount';
 
 export class MenuPage extends BasePage {
   constructor(page) {
     super(page);
     this._url = '/';
-    this.cartLink = page.getByLabel('Cart page');
-    this.totalCheckout = page.getByTestId('checkout');
-    this.promoMessage = page.getByText(
-      "It's your lucky day! Get an extra cup of Mocha for $4.",
-    );
-    this.yesPromoButton = page.getByRole('button', { name: 'Yes, of course!' });
-    this.noPromoButton = page.getByRole('button', { name: "Nah, I'll skip." });
-  }
-
-  coffeeCupLocator(coffeeName) {
-    const testId = coffeeName.replace(' ', '_');
-
-    return this.page.getByTestId(testId);
-  }
-
-  coffeeCupCostLocator(coffeeName) {
-    const coffeeCup = this.coffeeCupLocator(coffeeName);
-    return this.page.getByRole('listitem').filter({ has: coffeeCup });
-  }
-
-  async clickCoffeeCup(coffeeName) {
-    await step(`Click ${coffeeName} cup`, async () => {
-      await this.coffeeCupLocator(coffeeName).click();
-    });
-  }
-
-  async clickCartLink() {
-    await step(`Click 'Cart' link`, async () => {
-      await this.cartLink.click();
-    });
-  }
-
-  async clickYesPromoButton() {
-    await step(`Click 'Yes' promo button`, async () => {
-      await this.yesPromoButton.click();
-    });
-  }
-
-  async clickNoPromoButton() {
-    await step(`Click 'No' promo button`, async () => {
-      await this.noPromoButton.click();
-    });
-  }
-
-  async assertTotalCheckoutContainsValue(value) {
-    await step(`Assert Total checkout has value: ${value}`, async () => {
-      await expect(this.totalCheckout).toContainText(value);
-    });
-  }
-
-  async assertCoffeeCupCostHasValue(coffee, value) {
-    await step(`Assert ${coffee} cup cost has value: ${value}`, async () => {
-      await expect(this.coffeeCupCostLocator(coffee)).toContainText(value);
-    });
-  }
-
-  async assertPromoMessageIsVisible() {
-    await step(`Assert promo message is visible`, async () => {
-      await expect(this.promoMessage).toBeVisible();
-    });
+    this.header = new Header(page);
+    this.promo = new Promo(page);
+    this.cup = new Cup(page);
+    this.totalCount = new TotalCount(page);
   }
 }
