+++ tests/customerSite/cart/emptyCart/emptyCartShowsMessage.spec.js
@@ -3,5 +3,5 @@ import { test } from '../../../_fixtures/fixtures';
 test('An empty cart shows correct message', async ({ cartPage }) => {
   await cartPage.open();
 
-  await cartPage.assertNoCoffeeMessageIsVisible();
+  await cartPage.cartList.assertNoCoffeeMessageIsVisible();
 });
