+++ tests/customerSite/cart/refreshCart/cartCleanedAfterPageRefresh.spec.js
@@ -3,15 +3,15 @@ import { COFFEE_NAMES } from '../../../../src/constants';
 
 test('Cart cleaned after page refresh', async ({ cartPage, menuPage }) => {
   await menuPage.open();
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.cappuccino);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.cappuccino);
 
-  await menuPage.clickCartLink();
+  await menuPage.header.clickCartLink();
   await cartPage.waitForLoading();
 
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
 
   await cartPage.reload();
 
-  await cartPage.assertCoffeeItemIsHidden(COFFEE_NAMES.cappuccino);
-  await cartPage.assertNoCoffeeMessageIsVisible();
+  await cartPage.cartList.assertCoffeeItemIsHidden(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.assertNoCoffeeMessageIsVisible();
 });
