+++ tests/customerSite/cart/removeFromCart/coffeeRemovedFromCart.spec.js
@@ -13,12 +13,12 @@ testParameters.forEach(({ coffee }) => {
     menuPage,
   }) => {
     await menuPage.open();
-    await menuPage.clickCoffeeCup(coffee);
+    await menuPage.cup.clickCoffeeCup(coffee);
 
-    await menuPage.clickCartLink();
+    await menuPage.header.clickCartLink();
     await cartPage.waitForLoading();
 
-    await cartPage.clickCoffeeListItemRemoveAllButton(coffee);
-    await cartPage.assertNoCoffeeMessageIsVisible();
+    await cartPage.cartList.clickCoffeeListItemRemoveAllButton(coffee);
+    await cartPage.cartList.assertNoCoffeeMessageIsVisible();
   });
 });
