+++ tests/customerSite/cart/updateCart/cartUpdatedAfterClickingMinusButton.spec.js
@@ -6,21 +6,25 @@ test('Cart updated correctly after clicking minus for drinks', async ({
   menuPage,
 }) => {
   await menuPage.open();
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.cappuccino);
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.espresso);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.cappuccino);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.espresso);
 
-  await menuPage.clickCartLink();
+  await menuPage.header.clickCartLink();
   await cartPage.waitForLoading();
 
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.espresso);
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.espresso);
 
-  await cartPage.clickCoffeeListItemRemoveOneButton(COFFEE_NAMES.espresso);
+  await cartPage.cartList.clickCoffeeListItemRemoveOneButton(
+    COFFEE_NAMES.espresso,
+  );
 
-  await cartPage.assertCoffeeItemIsHidden(COFFEE_NAMES.espresso);
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.assertCoffeeItemIsHidden(COFFEE_NAMES.espresso);
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
 
-  await cartPage.clickCoffeeListItemRemoveOneButton(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.clickCoffeeListItemRemoveOneButton(
+    COFFEE_NAMES.cappuccino,
+  );
 
-  await cartPage.assertCoffeeItemIsHidden(COFFEE_NAMES.cappuccino);
-  await cartPage.assertNoCoffeeMessageIsVisible();
+  await cartPage.cartList.assertCoffeeItemIsHidden(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.assertNoCoffeeMessageIsVisible();
 });
