+++ tests/customerSite/promotions/mochaNotAddedToCartOnPromoDecline.spec.js
@@ -6,19 +6,19 @@ test('Discounted Mocha Not added to the Cart after promo rejecting', async ({
   menuPage,
 }) => {
   await menuPage.open();
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.cappuccino);
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.espresso);
-  await menuPage.clickCoffeeCup(COFFEE_NAMES.americano);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.cappuccino);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.espresso);
+  await menuPage.cup.clickCoffeeCup(COFFEE_NAMES.americano);
 
-  await menuPage.assertPromoMessageIsVisible();
-  await menuPage.clickNoPromoButton();
+  await menuPage.promo.assertPromoMessageIsVisible();
+  await menuPage.promo.clickNoPromoButton();
 
-  await menuPage.clickCartLink();
+  await menuPage.header.clickCartLink();
   await cartPage.waitForLoading();
 
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.espresso);
-  await cartPage.assertCoffeeItemIsHidden('(Discounted) Mocha');
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.espresso);
+  await cartPage.cartList.assertCoffeeItemIsHidden('(Discounted) Mocha');
 
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
-  await cartPage.assertCoffeeItemIsVisible(COFFEE_NAMES.americano);
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.cappuccino);
+  await cartPage.cartList.assertCoffeeItemIsVisible(COFFEE_NAMES.americano);
 });
